python-etl/requirements.txt
#install postgres connector
psycopg2-binary
ollama
requests
pdfplumber
scikit-learn
sqlalchemy
pandas